GI-Fix: Weltpositionen im GI-G-Buffer als fp32 (Ray-Origin-Praezision)
Der GI-G-Buffer speicherte Weltpositionen in RGBA16F. fp16 hat nur 10
Mantissenbits: bei Terrain-Koordinaten im Hunderterbereich betraegt die
Schrittweite schon ~0.125-0.25 Weltunits -- deutlich groeber als der
Normal-Offset von 0.05, mit dem gi_shadow.comp den Ray-Origin von der
Oberflaeche abhebt. Die Origins rasten damit auf ein Gitter ein und
landen teils unter der eigenen Oberflaeche: Die Shadow-Rays treffen ihr
eigenes Dreieck, was als fleckige/gebaenderte Schattenartefakte auf
grossen Landscape-Objekten sichtbar wird (nur mit GI an).

Position jetzt RGBA32F auf allen 5 Backends. Normalen bleiben fp16 --
sie sind einheitslang, da reicht die Praezision. Kosten: ein halbaufgeloester
RGBA32F-Buffer (bei 1080p ~8 MB statt 4 MB).

Formate muessen an mehreren Stellen zusammenpassen, sonst schlaegt die
Pipeline-/View-Erstellung fehl: Vulkan Render-Pass-Attachment + Image,
D3D11 Textur (Views erben via nullptr-Desc), D3D12 Resource + 3 SRVs +
PSO-RTVFormats[0], Metal Textur-Descriptor (jetzt getrennt von der
Normalen) + Pipeline-colorAttachments[0].

+++ src/HE_Rendering/src/Backends/OpenGL/OpenGLRenderer.cpp
@@ -4766,7 +4766,11 @@ void OpenGLRenderer::EnsureGiShadowTargets(int width, int height)
 	};
 
 	// World-space G-buffer: pos + normal MRT + depth.
-	m_giGBufPosTex  = makeTex(GL_RGBA16F, GL_NEAREST);
+	// Position must be fp32: it seeds the shadow-ray origins, and fp16 carries only a
+	// 10-bit mantissa (~0.25 world units at terrain-scale coordinates) — far coarser
+	// than the 0.05 normal offset, so origins quantise below the surface and the rays
+	// self-intersect in banded blotches. Normals stay fp16 (unit length).
+	m_giGBufPosTex  = makeTex(GL_RGBA32F, GL_NEAREST);
 	m_giGBufNormTex = makeTex(GL_RGBA16F, GL_NEAREST);
 	glGenTextures(1, &m_giGBufDepth);
 	glBindTexture(GL_TEXTURE_2D, m_giGBufDepth);
